requirements.txt
click==8.1.7
ollama==0.3.1
progressbar2==4.4.2
